Split Forms.js into two files, SignUpForm.js and LogInForm.js
